requirements.txt
coloredlogs==15.0.1
filelock==3.25.2
flatbuffers==25.12.19
fsspec==2026.2.0
humanfriendly==10.0
Jinja2==3.1.6
MarkupSafe==3.0.3
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
onnx==1.20.1
onnx-ir==0.2.0
onnxruntime==1.23.2
onnxscript==0.6.2
packaging==26.0
pillow==12.1.1
protobuf==7.34.1
sympy==1.14.0
torch==2.11.0
torchvision==0.26.0
typing_extensions==4.15.0
